Add on-the-fly optimizations in EmitterContext
Copy propagate, constant fold and simplify as we emit nodes in the
CFG in the EmitterContext.

This improves the CQ of LCQ by quite a bit. Reduces textual assembly
dump from 4.06 mb to 3.98 mb (258 dumps).

Will eventually have to measure its impact on throughput.

But the intuition for LCQ is that JIT time for small translation will
be increased and for larger ones decreased, since the register
allocator will have less variable to handle.